+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Ref: Disable `print()` breadcrumbs by default on Flutter
+
 # 6.0.0-beta.1
 
 * Feat: Browser detection (#502)